LineItemV2_int_to_long_fiх

See merge request qa/automag!1992

+++ web-tests/api/src/main/java/ru/instamart/api/dataprovider/RestDataProvider.java
@@ -549,7 +549,7 @@ public class RestDataProvider extends RestBase {
 
     @DataProvider(name = "changeLineItems")
     public static Object[][] changeLineItems() {
-        Integer productId = apiV2.dropAndFillCart(
+        Long productId = apiV2.dropAndFillCart(
                 SessionFactory.getSession(SessionType.API_V2).getUserData(),
                 EnvironmentProperties.DEFAULT_SID
         ).get(0).getId();
